iOS 11.2.6

+++ Frameworks/HealthKit.framework/HKWorkout.h
@@ -105,7 +105,7 @@
 
 - (id)_fiui_averagePaceQty;
 - (id)fiui_activityType;
-- (long long)fiui_averageCadence;
+- (double)fiui_averageCadence;
 - (double)fiui_averagePace;
 - (double)fiui_averagePower;
 - (double)fiui_averageTimePerLap;
