+++ PrivateFrameworks/SpotlightServices.framework/PRSModelFeedbackContext.h
@@ -6,24 +6,24 @@
     NSString * _l2ShadowVersion;
     NSString * _l2Version;
     NSString * _l3Version;
-    double * _scores;
+    float * _scores;
 }
 
 @property (nonatomic, retain) NSString *l2ShadowVersion;
 @property (nonatomic, retain) NSString *l2Version;
 @property (nonatomic, retain) NSString *l3Version;
-@property (nonatomic) double*scores;
+@property (nonatomic) float*scores;
 
 - (void).cxx_destruct;
 - (void)dealloc;
-- (id)initWithScores:(double*)arg1 models:(id)arg2 shadowModels:(id)arg3;
+- (id)initWithScores:(float*)arg1 models:(id)arg2 shadowModels:(id)arg3;
 - (id)l2ShadowVersion;
 - (id)l2Version;
 - (id)l3Version;
-- (double*)scores;
+- (float*)scores;
 - (void)setL2ShadowVersion:(id)arg1;
 - (void)setL2Version:(id)arg1;
 - (void)setL3Version:(id)arg1;
-- (void)setScores:(double*)arg1;
+- (void)setScores:(float*)arg1;
 
 @end
