+++ protocols/PRSFeatureVector.h
@@ -13,12 +13,12 @@
 - (NSArray *)getExpandedFeatureVectorAsArray;
 - (NSMutableDictionary *)getExpandedFeaturesAsDictionary;
 - (NSMutableDictionary *)getFeaturesAsDictionary;
-- (double)getServerFeatureAtIndex:(unsigned long long)arg1;
-- (double)getValueAtIndex:(unsigned long long)arg1;
-- (double)getValueForFeature:(NSString *)arg1;
+- (float)getServerFeatureAtIndex:(unsigned long long)arg1;
+- (float)getValueAtIndex:(unsigned long long)arg1;
+- (float)getValueForFeature:(NSString *)arg1;
 - (void)resetFeatureValues;
 - (void)setBundle_id:(NSString *)arg1;
-- (void)setExpandedSet:(double*)arg1;
+- (void)setExpandedSet:(float*)arg1;
 - (void)setValue:(NSNumber *)arg1 forFeature:(NSString *)arg2;
 - (void)setValue:(NSNumber *)arg1 forFeatureEnum:(unsigned long long)arg2;
